Rename `CustomOptionSupport` to `CustomOption`
